Generics added and jsp edited

+++ src/main/java/mikhail/shell/ems/models/Employee.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 public class Employee {
-    @Getter private long id;
+    @Getter private int id;
     @Getter @Setter private String name, surname, position, department;
     @Getter @Setter private byte age;
     public Employee () {}
